s/disk_log: made size_bytes_after_offset more accurate
Previously the `disk_log_impl::size_bytes_after_offset` was returning a
very rough approximation of the partition size after offset as it
limited the result accuracy to the segment sized. Changed the code to
return a better approximation of size after offset by using a segment
index. When requested offset is within the segment a closest segment
index entry is queried, the entry contains a position in segment file,
the position is an offset in bytes from the beginning of the file. By
subtracting the position from file size the code calculates the
remaining segment part with the accuracy of segment index entry interval
which by default is equal to 32KiB.

Signed-off-by: Michał Maślanka <michal@redpanda.com>