swarm.py: introduce base class
producer and client swarm services in DT are almost exact copy-paste
duplicates of each other (they use the same binary, in fact). They are
starting to diverge in annoying ways as bugs are fixed in one but not
the other, etc.

Introduce a base class to hold the common parts: the existing swarms
are then just small shells on top of it.