config: add gets_restored tag to properties
Cluster properties that _don't_ get restored during cluster restore were
previously managed in a static list in the cluster restore reconciler.
This made it easy to miss when adding new properties.

As a step towards making this list more future proof, this commit adds a
gets_restored flag to property metadata, and updates the reconciler to
generate its list dynamically based on the cluster's current properties.

The generated list is identical to the one previously defined in the
reconciler.